+++ CHANGELOG.md
@@ -3,6 +3,8 @@ so that developers know what code to modify, if any modifications are necessary.
 
 ### v0.1.3
 
+- Created RadioButtonWidget example: `radio_widget`
+
 ### v0.1.2
 
 [PR #5](https://github.com/KenSuenobu/pushrod/issues/5): Remove unnecessary borrowing from examples