+++ CHANGELOG.md
@@ -6,6 +6,8 @@ so that developers know what code to modify, if any modifications are necessary.
 - Created RadioButtonWidget example: `radio_widget`
 - Removed "get_by_name" as this function was not in use anywhere.
 
+[PR #9](https://github.com/KenSuenobu/pushrod/issues/9): Call build_layout on `Widget`s when appropriate
+
 ### v0.1.2
 
 [PR #5](https://github.com/KenSuenobu/pushrod/issues/5): Remove unnecessary borrowing from examples